requirements.txt
pyyaml
# pandas==1.4.3